# Use Unicode values instead of `HashMap` for performance