--- a/lindera/Cargo.toml
+++ b/lindera/Cargo.toml
@@ -26,7 +26,6 @@ anyhow = "1.0.68"
 bincode = "1.3.3"
 byteorder = "1.4.3"
 encoding = "0.2.33"
-once_cell = "1.16.0"
 regex = "1.7.0"
 serde = {version="1.0.151", features = ["derive"] }
 serde_json = "1.0.91"
@@ -48,6 +47,7 @@ lindera-unidic-builder = { version = "0.19.1", path = "../lindera-unidic-builder
 
 [dev-dependencies]
 criterion = { version = "0.4.0", features = ["html_reports"] }
+once_cell = "1.16.0"
 
 [[bench]]
 name = "bench"