I was using this library in a personal project and noticed a place where you can improve performance. Instead of using a static `HashMap<char, char>`, you can significantly improve performance by processing the character codes directly. Hopefully this suggestion is welcome.

A few notes:
- I noticed that the `KATAKANA_DAKUON_MAP` contains a single handakuten&rarr;dakuten conversion: パ&rarr;バ. Not sure if this was intentional, so I commented it out for now.
- There are some uncommon characters that are not covered: ゔ・ヴ, ヷ・ヸ・ヹ・ヺ. I'm assuming the exclusion of the latter four is intentional, but I wonder about the first two.

Output of `bench_hiragana_detection` test:
```text
new: 197ms 4000000
old: 4248ms 4000000
```

Output of `bench_hiragana_conversion` test:
```text
new: 129ms 4000000
old: 934ms 4000000
```