@mosuka 2022-12-24
Hi @encody ,
Thanks for using Lindera. And thanks for your great suggestions.
You are correct in pointing out that the conversion from 'パ' -> 'バ' is unnecessary.
Let us merge your code.
BTW, is this pull request still a work in progress?