@encody 2022-12-24
There are unnecessary pieces that I left in for testing purposes. Since it sounds like you would like to merge it, I will remove them.